# core-banking-accounts-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-accounts
  description: Banking Accounts Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/accounts
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/accounts/{accountId}/status-history/{historyId}:
    get:
      tags:
        - Account Status History
      summary: Get Account Status History by ID
      description: Retrieve a specific status history record by its unique identifier, ensuring it belongs to the specified account.
      operationId: getStatusHistory
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: historyId
          in: path
          description: Unique identifier of the status history record
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the account status history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatusHistoryDTO'
        '404':
          description: Status history record not found
    put:
      tags:
        - Account Status History
      summary: Update Account Status History
      description: Update an existing status history record associated with the specified account.
      operationId: updateStatusHistory
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: historyId
          in: path
          description: Unique identifier of the status history record to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountStatusHistoryDTO'
        required: true
      responses:
        '200':
          description: Status history record updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatusHistoryDTO'
        '404':
          description: Status history record not found
    delete:
      tags:
        - Account Status History
      summary: Delete Account Status History
      description: Remove an existing status history record by its unique identifier.
      operationId: deleteStatusHistory
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: historyId
          in: path
          description: Unique identifier of the status history record to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Status history record deleted successfully
        '404':
          description: Status history record not found
...
  /api/v1/accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Get Account by ID
      description: Retrieve an existing bank account by its unique identifier.
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
    put:
      tags:
        - Accounts
      summary: Update Account
      description: Update the details of an existing bank account by its unique identifier.
      operationId: updateAccount
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
    delete:
      tags:
        - Accounts
      summary: Delete Account
      description: Delete an existing bank account by its unique identifier.
      operationId: deleteAccount
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/accounts/{accountId}/status-history:
    get:
      tags:
        - Account Status History
      summary: List Account Status History
      description: Retrieve a paginated list of status history records associated with the specified account.
      operationId: getAllStatusHistory
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Successfully retrieved status history records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No status history records found for the specified account
    post:
      tags:
        - Account Status History
      summary: Create Account Status History
      description: Create a new status history record for the specified account.
      operationId: createStatusHistory
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountStatusHistoryDTO'
        required: true
      responses:
        '201':
          description: Status history record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatusHistoryDTO'
        '400':
          description: Invalid status history data provided
...
  /api/v1/account-restrictions/account/{accountId}:
    get:
      tags:
        - Account Restrictions
      summary: Get Account Restrictions by Account ID
      description: Retrieve all restrictions for a specific account.
      operationId: getAccountRestrictionsByAccountId
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the restrictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRestrictionDTO'
        '400':
          description: Invalid account ID
...
  /api/v1/account-notifications/account/{accountId}:
    get:
      tags:
        - Account Notifications
      summary: Get Account Notifications by Account ID
      description: Retrieve all notifications for a specific account.
      operationId: getAccountNotificationsByAccountId
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountNotificationDTO'
        '400':
          description: Invalid account ID
...
components:
  schemas:
...
    AccountStatusHistoryDTO:
      required:
        - accountId
        - statusCode
        - statusStartDatetime
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountStatusHistoryId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        statusCode:
          type: string
          enum:
            - OPEN
            - SUSPENDED
            - CLOSED
            - DORMANT
        statusStartDatetime:
          type: string
          format: date-time
        statusEndDatetime:
          type: string
          format: date-time
        reason:
          maxLength: 500
          minLength: 0
          type: string
...
    AccountDTO:
      required:
        - accountNumber
        - accountStatus
        - accountType
        - branchId
        - contractId
        - currency
        - openDate
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountId:
          type: string
          format: uuid
          readOnly: true
        contractId:
          type: string
          format: uuid
        accountNumber:
          maxLength: 50
          minLength: 0
          type: string
        accountType:
          type: string
          enum:
            - CHECKING
            - SAVINGS
            - TERM_DEPOSIT
            - LOAN
            - CREDIT_CARD
            - INVESTMENT
            - MORTGAGE
            - BUSINESS_CHECKING
            - BUSINESS_SAVINGS
            - FOREIGN_CURRENCY
            - PREMIUM
            - YOUTH
            - SENIOR
            - PENSION
            - GOVERNMENT_BENEFITS
            - CRYPTO_WALLET
            - TOKENIZED_ASSET
            - DIGITAL_ASSET
            - STABLECOIN
            - NFT_COLLECTION
        currency:
          maxLength: 3
          minLength: 3
          pattern: ^[A-Z]{3}$
          type: string
        openDate:
          type: string
          format: date
        closeDate:
          type: string
          format: date
        accountStatus:
          type: string
          enum:
            - OPEN
            - CLOSED
            - SUSPENDED
            - DORMANT
        branchId:
          type: string
          format: uuid
        description:
          maxLength: 500
          minLength: 0
          type: string
        accountSubType:
          type: string
          enum:
            - STANDARD_CHECKING
            - INTEREST_CHECKING
            - REWARDS_CHECKING
            - FREE_CHECKING
            - PREMIUM_CHECKING
            - STANDARD_SAVINGS
            - HIGH_YIELD_SAVINGS
            - MONEY_MARKET
            - CERTIFICATE_OF_DEPOSIT
            - IRA
            - FIXED_TERM_DEPOSIT
            - VARIABLE_TERM_DEPOSIT
            - PERSONAL_LOAN
            - AUTO_LOAN
            - HELOC
            - STANDARD_CREDIT_CARD
            - REWARDS_CREDIT_CARD
            - CASHBACK_CREDIT_CARD
            - TRAVEL_CREDIT_CARD
            - BUSINESS_CREDIT_CARD
            - FIXED_RATE_MORTGAGE
            - ADJUSTABLE_RATE_MORTGAGE
            - INTEREST_ONLY_MORTGAGE
            - STANDARD_BUSINESS_CHECKING
            - PREMIUM_BUSINESS_CHECKING
            - STANDARD_BUSINESS_SAVINGS
            - PREMIUM_BUSINESS_SAVINGS
            - MULTI_CURRENCY
            - SINGLE_CURRENCY_FOREIGN
            - RELATIONSHIP_PREMIUM
            - WEALTH_MANAGEMENT
            - YOUTH_SAVINGS
            - STUDENT_CHECKING
            - SENIOR_CHECKING
            - SENIOR_SAVINGS
            - PENSION_DISTRIBUTION
            - SOCIAL_SECURITY_BENEFITS
            - GOVERNMENT_ASSISTANCE
        taxReportingStatus:
          type: string
          enum:
            - REPORTABLE
            - EXEMPT
            - BACKUP_WITHHOLDING
            - FATCA_REPORTABLE
            - CRS_REPORTABLE
            - FATCA_CRS_REPORTABLE
            - PENDING_DOCUMENTATION
            - INCOMPLETE_INFORMATION
            - NON_REPORTABLE
        regulatoryStatus:
          type: string
          enum:
            - COMPLIANT
            - PENDING_DOCUMENTATION
            - ENHANCED_DUE_DILIGENCE
            - UNDER_REVIEW
            - SPECIAL_MONITORING
            - SANCTIONS_SCREENING
            - PEP_MONITORING
            - HIGH_RISK_MONITORING
            - REGULATORY_EXEMPTION
            - NON_COMPLIANT
            - REGULATORY_RESTRICTED
        maturityDate:
          type: string
          format: date
        interestAccrualMethod:
          type: string
          enum:
            - SIMPLE
            - DAILY_COMPOUND
            - MONTHLY_COMPOUND
            - QUARTERLY_COMPOUND
            - SEMI_ANNUAL_COMPOUND
            - ANNUAL_COMPOUND
            - CONTINUOUS_COMPOUND
            - DAILY_AVERAGE_BALANCE
            - MONTHLY_AVERAGE_BALANCE
            - MINIMUM_BALANCE
            - TIERED_RATE
            - STEPPED_RATE
            - DAY_COUNT_30_360
            - DAY_COUNT_ACTUAL_360
            - DAY_COUNT_ACTUAL_365
            - DAY_COUNT_ACTUAL_ACTUAL
        interestPaymentFrequency:
          type: string
          enum:
            - DAILY
            - WEEKLY
            - BI_WEEKLY
            - MONTHLY
            - QUARTERLY
            - SEMI_ANNUALLY
            - ANNUALLY
            - AT_MATURITY
            - AT_CLOSURE
            - SPECIFIC_DAY_OF_MONTH
            - SPECIFIC_DATE
            - NONE
        minimumBalance:
          minimum: 0
          exclusiveMinimum: false
          type: number
        overdraftLimit:
          minimum: 0
          exclusiveMinimum: false
          type: number
        walletAddress:
          maxLength: 100
          minLength: 0
          type: string
        blockchainNetwork:
          maxLength: 50
          minLength: 0
          type: string
        tokenContractAddress:
          maxLength: 100
          minLength: 0
          type: string
        tokenStandard:
          maxLength: 20
          minLength: 0
          type: string
        isCustodial:
          type: boolean
      description: Data for the new account
...
    AccountRestrictionDTO:
      required:
        - accountId
        - appliedBy
        - reason
        - restrictionType
        - startDateTime
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountRestrictionId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        restrictionType:
          type: string
          enum:
            - WITHDRAWAL_HOLD
            - DEPOSIT_HOLD
            - ACCOUNT_FREEZE
            - LEGAL_ORDER
            - COURT_ORDER
            - FRAUD_INVESTIGATION
            - DECEASED_HOLDER
            - DORMANCY
            - NEGATIVE_BALANCE
            - EXCESSIVE_OVERDRAFTS
            - SUSPICIOUS_ACTIVITY
            - AML_INVESTIGATION
            - INTERNATIONAL_TRANSFER_RESTRICTION
            - HIGH_VALUE_TRANSACTION_RESTRICTION
            - PENDING_DOCUMENTATION
            - EXPIRED_IDENTIFICATION
            - BANKRUPTCY
            - ESTATE_SETTLEMENT
            - REGULATORY_COMPLIANCE
            - SANCTIONS_SCREENING
            - TRANSACTION_TYPE_RESTRICTION
            - CHANNEL_RESTRICTION
            - GEOGRAPHIC_RESTRICTION
            - CLOSURE_IN_PROGRESS
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        restrictedAmount:
          minimum: 0
          exclusiveMinimum: false
          type: number
        referenceNumber:
          maxLength: 100
          minLength: 0
          type: string
        reason:
          maxLength: 500
          minLength: 0
          type: string
        appliedBy:
          maxLength: 100
          minLength: 0
          type: string
        removedBy:
          maxLength: 100
          minLength: 0
          type: string
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
    AccountNotificationDTO:
      required:
        - accountId
        - creationDateTime
        - message
        - notificationType
        - title
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountNotificationId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        notificationType:
          type: string
          enum:
            - LOW_BALANCE
            - LARGE_DEPOSIT
            - LARGE_WITHDRAWAL
            - OVERDRAFT
            - INSUFFICIENT_FUNDS
            - APPROACHING_MINIMUM_BALANCE
            - STATEMENT_AVAILABLE
            - INTEREST_PAYMENT
            - FEE_CHARGED
            - STATUS_CHANGE
            - APPROACHING_MATURITY
            - MATURITY_REACHED
            - SUSPICIOUS_ACTIVITY
            - RESTRICTION_APPLIED
            - RESTRICTION_REMOVED
            - APPROACHING_DORMANCY
            - DORMANCY_APPLIED
            - GOAL_PROGRESS
            - GOAL_ACHIEVED
            - AUTOMATIC_TRANSFER
            - AUTOMATIC_TRANSFER_FAILED
            - BALANCE_THRESHOLD
            - PARAMETER_CHANGE
            - ACCOUNT_CLOSURE
        title:
          maxLength: 200
          minLength: 0
          type: string
        message:
          maxLength: 1000
          minLength: 0
          type: string
        creationDateTime:
          type: string
          format: date-time
        expiryDateTime:
          type: string
          format: date-time
        isRead:
          type: boolean
        readDateTime:
          type: string
          format: date-time
        priority:
          maximum: 10
          minimum: 1
          type: integer
          format: int32
        deliveryChannels:
          maxLength: 100
          minLength: 0
          type: string
        eventReference:
          maxLength: 100
          minLength: 0
          type: string
        relatedAmount:
          minimum: 0
          exclusiveMinimum: false
          type: number
        actionUrl:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://.*)?$
          type: string
        actionText:
          maxLength: 100
          minLength: 0
          type: string
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...